Move SourceChainJsonDump to holochain_types
